modularize the client code into modules to make management easier
